split proxy API
